Keep buffer in_use set for longer
We were clearing the buffer's in_use field too early: in
atomic_flip_handler, KMS is notifying us that the hardware has now
started displaying that buffer. Split the output's buffer tracking into
two buffers: one for a 'pending' buffer and one for a 'last' buffer.

The pending buffer is the one we have committed to KMS in
repaint_one_output and the main loop, but has not started displaying
until we receive the event in atomic_flip_handler. When we get the
event, move it to buffer_last, and free the old buffer_last, which is
now no longer being displayed by the hardware, since it has been
replaced by the next.

This fixes tearing seen by starting to render into the same buffer as
it's being displayed.