Add optional logind support to manage sessions
This commit adds basic logind support to initiate a session and to
take/release the KMS device. VT switching is not implemented as it
doesn't seem to add much value to the purpose of this sample.

Most of the code has been copied and adapted from wlroots' logind
session backend. Same as in wlroots, logind can be provided by systemd
or elogind.

As a result, VTs are restored much cleaner and more predictable when
running the sample.

logind is automatically enabled if the appropriate libraries are
present on the system.

The direct method is still supported if logind support can't be enabled.