Add optional keyboard support to quit sample
This commit adds very basic keyboard support to poll whether the ESC key
was pressed. If it was, the sample quits gracefully. This is useful when
running the sample from a single machine without remote access to the
machine to kill the process.

The feature is enabled if libudev and libinput is found during configure
time.

If logind support is enabled, input devices are opened via logind rather
than directly.