Fix comment about the OUT_FENCE timestamp
The timestamp of the KMS commit out-fence is _not_ the same as the
completion timestamp passed to the event handler. Many drivers support
"high-precision" timestamps, where the timestamp passed to the handler
is corrected by the hardware to some implementation-dependent point in
the vblank sequence. The timestamp of the out-fence, however, is closer
to the time of the display hardware interrupt.

Simply drop the comment there for now.