main: rename 'completion' to 'event_time'
Rename the 'completion' timestamp to 'event_time'. This is more vague,
but unfortunately better representing the current state. The name
'completion' would suggest a timestamp in the past, i.e. when the commit
became active, which is most often not the case for drivers with
high-precision timestamp support (see previous commit).